pyproject.toml
[tool.poetry]
name = "agentic-claims"
version = "0.1.0"
description = "Multi-agent multimodal expense claims processing system"
authors = []
packages = [{include = "agentic_claims", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.15"
langgraph = ">=0.4"
langgraph-checkpoint-postgres = ">=2.0"
chainlit = ">=2.0"
pydantic-settings = ">=2.0"
psycopg = {version = ">=3.0", extras = ["binary"]}
python-dotenv = ">=1.0"
alembic = ">=1.18"
sqlalchemy = {version = ">=2.0", extras = ["asyncio"]}
openai = ">=1.0"
langchain-openai = ">=0.3"
fastapi = ">=0.115"
uvicorn = {version = ">=0.34", extras = ["standard"]}
python-multipart = ">=0.0.20"
opencv-python-headless = ">=4.0"
mcp = ">=1.8"
httpx = ">=0.27"
numpy = ">=1.26"
python-json-logger = "^4.0.0"
langchain-openrouter = "^0.2.0"
openrouter = "0.7.11"
itsdangerous = "^2.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
pytest-asyncio = ">=0.24"
pytest-cov = ">=6.0"
ruff = ">=0.8"
pytailwindcss = ">=0.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "e2e: marks tests as end-to-end (requires running Docker services and OpenRouter API key)",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["N802", "N803", "N806", "N812", "N816", "N999"]
